refactor: Unify records validators in one single module
