+++ src/argilla_server/bulk/records_bulk.py
@@ -28,18 +28,17 @@ from argilla_server.contexts.records import (
     fetch_records_by_ids_as_dict,
 )
 from argilla_server.models import Dataset, Question, Record, Response, Suggestion, Vector, VectorSettings
-from argilla_server.schemas.v1.records import RecordCreate
+from argilla_server.schemas.v1.records import RecordCreate, RecordUpsert
 from argilla_server.schemas.v1.records_bulk import (
     RecordsBulk,
     RecordsBulkCreate,
     RecordsBulkUpsert,
     RecordsBulkWithUpdateInfo,
-    RecordUpsert,
 )
 from argilla_server.schemas.v1.responses import UserResponseCreate
 from argilla_server.schemas.v1.suggestions import SuggestionCreate
 from argilla_server.search_engine import SearchEngine
-from argilla_server.validators.records_bulk import RecordsBulkCreateValidator, RecordsBulkUpsertValidator
+from argilla_server.validators.records import RecordsBulkCreateValidator, RecordsBulkUpsertValidator
 from argilla_server.validators.responses import ResponseCreateValidator
 from argilla_server.validators.suggestions import SuggestionCreateValidator
 from argilla_server.validators.vectors import VectorValidator
